Fixed the floating point errors in GenericOptic.collect_rays.

+++ xicsrt/xics_rt_raytrace.py
@@ -115,7 +115,7 @@ def raytrace(source, detector, *optics, number_of_runs=None, collect_optics=None
 
         # Save only a portion of the lost rays so that our lost history does
         # not become too large.
-        max_lost = 100000/number_of_runs
+        max_lost = int(100000/number_of_runs)
         lost_max = min(max_lost, len(w_lost))
         index_lost = np.arange(len(w_lost))
         np.random.shuffle(index_lost)
